pyproject.toml
[tool.poetry]
name = "nonebot-plugin-copywrite"
version = "0.1.0"
description = ""
authors = ["Yan <1964649083@qq.com>"]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
nonebot2 = "^2.3.1"
nonebot-adapter-onebot = "^2.4.3"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
